:recycle: More one file one export refactoring
