+++ packages/block/src/blockFromRpc.ts
@@ -1,7 +1,7 @@
 import { TransactionFactory } from '@tevm/tx'
-import { type Hex, TypeOutput, hexToBytes, setLengthLeft, toBytes, toType } from '@tevm/utils'
+import { type Hex, hexToBytes } from '@tevm/utils'
 
-import { blockHeaderFromRpc } from './header-from-rpc.js'
+import { blockHeaderFromRpc } from './blockHeaderFromRpc.js'
 
 import { Block } from './index.js'
 
@@ -11,24 +11,7 @@ import { createClRequest } from './createClRequest.js'
 import type { JsonRpcBlock } from './JsonRpcBlock.js'
 import type { BlockOptions } from './BlockOptions.js'
 import type { BlockData } from './BlockData.js'
-
-function normalizeTxParams(_txParams: any) {
-	const txParams = Object.assign({}, _txParams)
-
-	txParams.gasLimit = toType(txParams.gasLimit ?? txParams.gas, TypeOutput.BigInt)
-	txParams.data = txParams.data === undefined ? txParams.input : txParams.data
-
-	// check and convert gasPrice and value params
-	txParams.gasPrice = txParams.gasPrice !== undefined ? BigInt(txParams.gasPrice) : undefined
-	txParams.value = txParams.value !== undefined ? BigInt(txParams.value) : undefined
-
-	// strict byte length checking
-	txParams.to = txParams.to !== null && txParams.to !== undefined ? setLengthLeft(toBytes(txParams.to), 20) : null
-
-	txParams.v = toType(txParams.v, TypeOutput.BigInt)
-
-	return txParams
-}
+import { normalizeTxParams } from './normalizeTxParams.js'
 
 /**
  * Creates a new block object from Ethereum JSON RPC.
