+++ packages/block/src/executionPayloadFromBeaconPayload.ts
@@ -1,91 +1,8 @@
-import { type Hex, numberToHex } from '@tevm/utils'
-import type { VerkleExecutionWitness } from './VerkleExecutionWitness.js'
+import { numberToHex } from '@tevm/utils'
 import type { ExecutionPayload } from './ExecutionPayload.js'
-
-type BeaconWithdrawal = {
-	index: Hex
-	validator_index: Hex
-	address: Hex
-	amount: Hex
-}
-
-// Payload json that one gets using the beacon apis
-// curl localhost:5052/eth/v2/beacon/blocks/56610 | jq .data.message.body.execution_payload
-export type BeaconPayloadJson = {
-	parent_hash: Hex
-	fee_recipient: Hex
-	state_root: Hex
-	receipts_root: Hex
-	logs_bloom: Hex
-	prev_randao: Hex
-	block_number: Hex
-	gas_limit: Hex
-	gas_used: Hex
-	timestamp: Hex
-	extra_data: Hex
-	base_fee_per_gas: Hex
-	block_hash: Hex
-	transactions: Hex[]
-	withdrawals?: BeaconWithdrawal[]
-	blob_gas_used?: Hex
-	excess_blob_gas?: Hex
-	parent_beacon_block_root?: Hex
-	// the casing of VerkleExecutionWitness remains same camel case for now
-	execution_witness?: VerkleExecutionWitness
-}
-
-type VerkleProofSnakeJson = {
-	commitments_by_path: Hex[]
-	d: Hex
-	depth_extension_present: Hex
-	ipa_proof: {
-		cl: Hex[]
-		cr: Hex[]
-		final_evaluation: Hex
-	}
-	other_stems: Hex[]
-}
-
-type VerkleStateDiffSnakeJson = {
-	stem: Hex
-	suffix_diffs: {
-		current_value: Hex | null
-		new_value: Hex | null
-		suffix: number | string
-	}[]
-}
-
-type VerkleExecutionWitnessSnakeJson = {
-	state_diff: VerkleStateDiffSnakeJson[]
-	verkle_proof: VerkleProofSnakeJson
-}
-
-function parseExecutionWitnessFromSnakeJson({
-	state_diff,
-	verkle_proof,
-}: VerkleExecutionWitnessSnakeJson): VerkleExecutionWitness {
-	return {
-		stateDiff: state_diff.map(({ stem, suffix_diffs }) => ({
-			stem,
-			suffixDiffs: suffix_diffs.map(({ current_value, new_value, suffix }) => ({
-				currentValue: current_value,
-				newValue: new_value,
-				suffix,
-			})),
-		})),
-		verkleProof: {
-			commitmentsByPath: verkle_proof.commitments_by_path,
-			d: verkle_proof.d,
-			depthExtensionPresent: verkle_proof.depth_extension_present,
-			ipaProof: {
-				cl: verkle_proof.ipa_proof.cl,
-				cr: verkle_proof.ipa_proof.cr,
-				finalEvaluation: verkle_proof.ipa_proof.final_evaluation,
-			},
-			otherStems: verkle_proof.other_stems,
-		},
-	}
-}
+import type { BeaconPayloadJson } from './BeaconPayloadJson.js'
+import type { VerkleExecutionWitnessSnakeJson } from './VerkleExecutionWitnessSnakeJson.js'
+import { parseExecutionWitnessFromSnakeJson } from './parseExecutionWitnessFromSnakeJson.js'
 
 /**
  * Converts a beacon block execution payload JSON object {@link BeaconPayloadJson} to the {@link ExecutionPayload} data needed to construct a {@link Block}.
