+++ packages/block/src/index.ts
@@ -1,6 +1,6 @@
-export { Block } from './block.js'
-export { type BeaconPayloadJson, executionPayloadFromBeaconPayload } from './from-beacon-payload.js'
-export { BlockHeader } from './header.js'
-export { blockFromRpc } from './from-rpc.js'
+export { Block } from './Block.js'
+export { executionPayloadFromBeaconPayload } from './executionPayloadFromBeaconPayload.js'
+export { BlockHeader } from './Header.js'
+export { blockFromRpc } from './blockFromRpc.js'
 export type { ClRequest } from './ClRequest.js'
 export * from './WithdrawalsBytes.js'
